Reworked to allow floating point values, with consolidation of splitter.py into mapper.py and factoring out treemap.py as main program for command-line interface.
